# src/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "sayori",
    "url" : "https://acow-api-dev.sayori.link"
  } ],
  "info" : {
    "description" : "This is a simple API",
    "version" : "1.0.0",
    "title" : "ACOW API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths" : {
...
    "/seasons/{seasonId}/signup" : {
      "parameters" : [ {
        "schema" : {
          "type" : "string"
        },
        "name" : "seasonId",
        "in" : "path",
        "required" : true
      } ],
      "get" : {
        "summary" : "Retrieve current user's signup status",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "signedUp" : {
                      "type" : "boolean"
                    }
                  },
                  "x-examples" : {
                    "Example 1" : {
                      "success" : true,
                      "signedUp" : true
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "signedUp" : true
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "get-seasons-seasonid-signup",
        "description" : "Duplicate operations are idempotent."
      },
      "post" : {
        "summary" : "Sign up for the season",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "Duplicate operations are idempotent.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericSuccessResponse"
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "post-seasons-seasonid-signup",
        "description" : "Duplicate operations are idempotent."
      },
      "delete" : {
        "summary" : "Remove signup for the season",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericSuccessResponse"
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "delete-seasons-seasonid-signup",
        "description" : "Duplicate operations are idempotent."
      }
    },
    "/seasons/{seasonId}/signups" : {
      "parameters" : [ {
        "schema" : {
          "type" : "string"
        },
        "name" : "seasonId",
        "in" : "path",
        "required" : true
      } ],
      "get" : {
        "summary" : "[Admin] Retrieve the season's signup list",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-examples" : {
                    "Example 1" : {
                      "success" : true,
                      "signups" : [ ]
                    }
                  },
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "signups" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "user" : {
                            "$ref" : "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "signups" : [ {
                        "user" : {
                          "id" : 1,
                          "username" : "user1"
                        }
                      } ]
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "get-seasons-seasonid-signups"
      }
    },
    "/seasons/{seasonId}/contract-types" : {
      "parameters" : [ {
        "schema" : {
          "type" : "string"
        },
        "name" : "seasonId",
        "in" : "path",
        "required" : true,
        "description" : "season id"
      } ],
      "get" : {
        "summary" : "Retrieves a season's available contract types",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "x-stoplight" : {
                        "id" : "2oyno6x3d4ijb"
                      }
                    },
                    "contractTypes" : {
                      "type" : "array",
                      "x-stoplight" : {
                        "id" : "s8akyjivw0wzo"
                      },
                      "items" : {
                        "$ref" : "#/components/schemas/ContractType"
                      }
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "contractTypes" : [ {
                        "id" : "1",
                        "name" : "Base Contract",
                        "slug" : "base",
                        "icon" : "kirby",
                        "discordChannelId" : "1077741836749242468",
                        "assignmentStart" : "2026-05-15T00:00:00.000Z",
                        "assignmentEnd" : "2026-05-16T00:00:00.000Z",
                        "reviewDeadline" : "2026-05-23T00:00:00.000Z"
                      } ]
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "get-seasons-seasonId-contract-types"
      },
      "post" : {
        "summary" : "[Admin] Create a new contract type for a season",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "x-stoplight" : {
                        "id" : "zd4he2whpi8t1"
                      }
                    },
                    "contractType" : {
                      "$ref" : "#/components/schemas/ContractType"
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "contractType" : {
                        "id" : "1",
                        "name" : "Dating Contract",
                        "slug" : "dating",
                        "icon" : "hearts",
                        "discordChannelId" : "1077741836749242468",
                        "assignmentStart" : "2026-05-15T00:00:00.000Z",
                        "assignmentEnd" : "2026-05-16T00:00:00.000Z",
                        "reviewDeadline" : "2026-05-23T00:00:00.000Z",
                        "createdAt" : "2026-05-14T00:00:00.000Z",
                        "updatedAt" : "2026-05-14T00:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "post-seasons-seasonId-contract-types",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContractTypeEditable"
              },
              "examples" : {
                "Example 1" : {
                  "value" : {
                    "name" : "Base Contract",
                    "slug" : "base",
                    "icon" : "kirby",
                    "discordChannelId" : "1077741836749242468",
                    "assignmentStart" : "2026-05-15T00:00:00.000Z",
                    "assignmentEnd" : "2026-05-16T00:00:00.000Z",
                    "reviewDeadline" : "2026-05-23T00:00:00.000Z"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/seasons/{seasonId}/contract-types/{slug}/contracts" : {
      "parameters" : [ {
        "schema" : {
          "type" : "string"
        },
        "name" : "seasonId",
        "in" : "path",
        "required" : true,
        "description" : "season id, accepts \"current\""
      }, {
        "schema" : {
          "type" : "string"
        },
        "name" : "slug",
        "in" : "path",
        "required" : true,
        "description" : "contract type slug, accepts numeric id"
      } ],
      "post" : {
        "summary" : "[Admin] Create contract",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "x-stoplight" : {
                        "id" : "ottap0auioo3y"
                      }
                    },
                    "contract" : {
                      "$ref" : "#/components/schemas/Contract"
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "contract" : {
                        "id" : "5",
                        "season" : "1",
                        "contractor" : {
                          "id" : "1",
                          "username" : "admin1",
                          "roles" : [ "Array" ],
                          "createdAt" : "2026-05-16T19:57:23.000Z",
                          "updatedAt" : "2026-05-16T19:57:23.000Z"
                        },
                        "contractee" : {
                          "id" : "6",
                          "username" : "user6",
                          "roles" : [ ],
                          "createdAt" : "2026-05-16T19:57:23.000Z",
                          "updatedAt" : "2026-05-16T19:57:23.000Z"
                        },
                        "contractType" : {
                          "id" : "1",
                          "season" : "1",
                          "name" : "Base Contract",
                          "slug" : "base",
                          "icon" : "kirby",
                          "discordChannelId" : "1077741836749242468",
                          "assignmentStart" : "2026-05-15T00:00:00.000Z",
                          "assignmentEnd" : "2026-05-16T00:00:00.000Z",
                          "reviewDeadline" : "2026-05-23T00:00:00.000Z",
                          "createdAt" : "2026-05-14T00:00:00.000Z",
                          "updatedAt" : "2026-05-14T00:00:00.000Z"
                        },
                        "name" : "",
                        "progress" : "",
                        "score" : "",
                        "reviewContent" : "",
                        "verdict" : "PENDING",
                        "createdAt" : "2026-05-14T00:01:00.001Z",
                        "updatedAt" : "2026-05-14T00:01:00.001Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "post-seasons-seasonId-contract-types-slug-contracts",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContractPostable"
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "User" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "roles" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "createdAt" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string"
          }
        },
        "x-examples" : {
          "Example 1" : {
            "id" : "308152268249628672",
            "username" : "_ganyuwu",
            "roles" : [ "admin" ],
            "createdAt" : "2026-05-07T19:31:56.000Z",
            "updatedAt" : "2026-05-07T19:31:56.000Z"
          }
        }
      },
...
      "GenericSuccessResponse" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "success" : false,
            "error" : "Cannot create a new season while there is an active season"
          }
        },
        "title" : "GenericSuccessResponse",
        "required" : [ "success" ],
        "properties" : {
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "ContractType" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "id" : "1",
            "name" : "Base Contract",
            "slug" : "base",
            "icon" : "kirby",
            "discordChannelId" : "1077741836749242468",
            "assignmentStart" : "2026-05-15T00:00:00.000Z",
            "assignmentEnd" : "2026-05-16T00:00:00.000Z",
            "reviewDeadline" : "2026-05-23T00:00:00.000Z"
          }
        },
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          },
          "discordChannelId" : {
            "type" : "string"
          },
          "assignmentStart" : {
            "type" : "string"
          },
          "assignmentEnd" : {
            "type" : "string"
          },
          "reviewDeadline" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "5wz0tdmqtnlqm"
            }
          },
          "createdAt" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "xl8m59bagv5xb"
            }
          },
          "updatedAt" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "wz0hc4x7ivdub"
            }
          }
        }
      },
      "ContractTypeEditable" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "id" : "1",
            "name" : "Base Contract",
            "slug" : "base",
            "icon" : "kirby",
            "discordChannelId" : "1077741836749242468",
            "assignmentStart" : "2026-05-15T00:00:00.000Z",
            "assignmentEnd" : "2026-05-16T00:00:00.000Z",
            "reviewDeadline" : "2026-05-23T00:00:00.000Z"
          }
        },
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          },
          "discordChannelId" : {
            "type" : "string"
          },
          "reviewDeadline" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "dvtfcmtmidoti"
            }
          },
          "assignmentStart" : {
            "type" : "string"
          },
          "assignmentEnd" : {
            "type" : "string"
          }
        }
      },
      "ContractPostable" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "contractor" : "4",
            "contractee" : "2"
          }
        },
        "properties" : {
          "contractor" : {
            "type" : "string"
          },
          "contractee" : {
            "type" : "string"
          }
        }
      },
      "Contract" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "id" : "5",
            "season" : "1",
            "contractor" : {
              "id" : "1",
              "username" : "admin1",
              "roles" : [ "Array" ],
              "createdAt" : "2026-05-16T19:57:23.000Z",
              "updatedAt" : "2026-05-16T19:57:23.000Z"
            },
            "contractee" : {
              "id" : "6",
              "username" : "user6",
              "roles" : [ ],
              "createdAt" : "2026-05-16T19:57:23.000Z",
              "updatedAt" : "2026-05-16T19:57:23.000Z"
            },
            "contractType" : {
              "id" : "1",
              "season" : "1",
              "name" : "Base Contract",
              "slug" : "base",
              "icon" : "kirby",
              "discordChannelId" : "1077741836749242468",
              "assignmentStart" : "2026-05-15T00:00:00.000Z",
              "assignmentEnd" : "2026-05-16T00:00:00.000Z",
              "reviewDeadline" : "2026-05-23T00:00:00.000Z",
              "createdAt" : "2026-05-14T00:00:00.000Z",
              "updatedAt" : "2026-05-14T00:00:00.000Z"
            },
            "name" : "",
            "progress" : "",
            "score" : "",
            "reviewContent" : "",
            "verdict" : "PENDING",
            "createdAt" : "2026-05-14T00:01:00.001Z",
            "updatedAt" : "2026-05-14T00:01:00.001Z"
          }
        },
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "season" : {
            "type" : "string"
          },
          "contractor" : {
            "$ref" : "#/components/schemas/User"
          },
          "contractee" : {
            "$ref" : "#/components/schemas/User"
          },
          "contractType" : {
            "$ref" : "#/components/schemas/ContractType"
          },
          "name" : {
            "type" : "string"
          },
          "progress" : {
            "type" : "string"
          },
          "score" : {
            "type" : "string"
          },
          "reviewContent" : {
            "type" : "string"
          },
          "verdict" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "access_token" : {
        "name" : "Authorization",
        "type" : "apiKey",
        "in" : "header"
      }
    }
  },
  "security" : [ {
    "access_token" : [ ]
  } ]
}